fix(cli): fix type mismatch in codegen block_size parameter
with_block_size expects u32, not usize. Removes unnecessary cast.

+++ crates/ringkernel-cli/src/commands/codegen.rs
@@ -225,7 +225,7 @@ fn generate_cuda_kernel(kernel: &KernelInfo) -> CliResult<String> {
         use ringkernel_cuda_codegen::{transpile_ring_kernel, RingKernelConfig};
 
         let config =
-            RingKernelConfig::new(&kernel.name).with_block_size(kernel.block_size as usize);
+            RingKernelConfig::new(&kernel.name).with_block_size(kernel.block_size);
 
         match transpile_ring_kernel(&kernel.function, &config) {
             Ok(code) => return Ok(code),
